⚡️ :: RankPrevieTableViewCell
